test(animation): migrate build and basic tests
